% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSUBstats-package.R
\docType{package}
\name{CSUBstats-package}
\alias{CSUBstats}
\alias{CSUBstats-package}
\title{CSUBstats: Statistical Functions and Data for Teaching}
\description{
Functions and case-study datasets for learning and applying selected statistical
methods. The package supports the accompanying training modules and provides
helpers for group comparisons, regression, and exploratory data analysis.
}
\seealso{
Useful starting points are \code{\link[=two.mean.test]{two.mean.test()}}, \code{\link[=sfaov]{sfaov()}}, \code{\link[=normqqplot]{normqqplot()}},
and the \link{motidf} dataset.
}
\author{
\strong{Maintainer}: Eduardo Montoya \email{emontoya2@csub.edu}

Authors:
\itemize{
  \item Eduardo Montoya \email{emontoya2@csub.edu}
}

}
\keyword{internal}
